+++ CIS425/Hw04/utils.h
@@ -8,9 +8,9 @@
 #endif
 
 #define PI 3.14159265358979324
-float degToRad(int angInDeg);
-void incAng(int& angInDeg);
-void decAng(int& angInDeg);
+float degToRad(double angInDeg);
+void incAng(double& angInDeg);
+void decAng(double& angInDeg);
 float cosDiffAng(float u[], float v[], int length);
 void delay(int sec);
